+++ scarlib-core/src/main/scala/it/unibo/scarlib/core/model/State.scala
@@ -4,29 +4,27 @@
  * listed, for each module, in the respective subproject's build.gradle.kts file.
  *
  * This file is part of ScaRLib, and is distributed under the terms of the
- * GNU General Public License as described in the file LICENSE in the ScaRLin distribution's top directory.
+ * MIT License as described in the file LICENSE in the ScaRLib distribution's top directory.
  */
 
 package it.unibo.scarlib.core.model
 
+import it.unibo.scarlib.core.neuralnetwork.NeuralNetworkEncoding
+
 /** A generic state in which the environment can be */
 trait State {
-
-    /** Gets the number of elements in the state */
-    def elements(): Int
-
-    /** Converts the state into a format usable by the neural network */
-    def toSeq(): Seq[Double]
-
-    /** Checks if the state is empty */
-    def isEmpty(): Boolean
+  def isEmpty(): Boolean
 }
 
 /** An empty state */
 class EmptyState extends State {
-    override def elements(): Int = 0
+  override def isEmpty(): Boolean = true
+}
 
-    override def toSeq(): Seq[Double] = Seq()
+object EmptyState {
+  implicit val encoding: NeuralNetworkEncoding[State] = new NeuralNetworkEncoding[State] {
+      override def elements(): Int = 0
 
-    override def isEmpty(): Boolean = true
+      override def toSeq(element: State): Seq[Double] = Seq.empty[Double]
+  }
 }