+++ scarlib-core/src/main/scala/it/unibo/scarlib/core/neuralnetwork/NeuralNetworkEncoding.scala
@@ -4,13 +4,15 @@
  * listed, for each module, in the respective subproject's build.gradle.kts file.
  *
  * This file is part of ScaRLib, and is distributed under the terms of the
- * GNU General Public License as described in the file LICENSE in the ScaRLin distribution's top directory.
+ * MIT License as described in the file LICENSE in the ScaRLib distribution's top directory.
  */
 
 package it.unibo.scarlib.core.neuralnetwork
 
-trait NeuralNetworkEncoding[A]{
+import it.unibo.scarlib.core.model.State
+
+trait NeuralNetworkEncoding[A <: State]{
     def elements(): Int
 
-    def toSeq(elem: A): Seq[Double]
+    def toSeq(element: A): Seq[Double]
 }