+++ vmas/src/main/scala/it/unibo/scarlib/vmas/Main.scala
@@ -1,14 +1,16 @@
 package it.unibo.scarlib.vmas
 
-import it.unibo.scarlib.core.deepRL.CTDESystem
+import it.unibo.scarlib.core.model.EmptyState.encoding
+import it.unibo.scarlib.core.system.CTDESystem
 import it.unibo.scarlib.core.model.{Action, DeepQLearner, LearningConfiguration, ReplayBuffer, State}
+import it.unibo.scarlib.core.neuralnetwork.NeuralNetworkEncoding
 import it.unibo.scarlib.core.util.Logger
 
 import scala.concurrent.ExecutionContext
 
-class Main extends App {
-
-    val stateDescriptor = VMASStateDescriptor(lidars = Seq[Int](15,15))
+object Main extends App {
+    val stateDescriptor = VMASStateDescriptor(hasPosition=false, hasVelocity=false, lidars = Seq[Int](50))
+    VMASState.setDescriptor(stateDescriptor)
     private val memory = ReplayBuffer[VMASState, VMASAction](10000)
     private val actions = VMASAction.toSeq
     private val learningConfiguration = new LearningConfiguration(
@@ -17,9 +19,10 @@ class Main extends App {
     )
     val dql = new DeepQLearner(memory.asInstanceOf[ReplayBuffer[State, Action]], actions, learningConfiguration)
     val rewardFunction = new CohisionAndCollisionRewardFunction()
-    private val envSettings = new VmasSettings(scenario = "CohisionAndCollision", nEnv = 2, nAgents = 4, nTargets = 0, nSteps = 100)
-    val environment = new VmasEnvironment(rewardFunction, actions, envSettings, WANDBLogger.getClass.asInstanceOf[Class[_ <: Logger]])
-    val nAgents = 4
+    val nAgents = 1
+    private val envSettings = VmasSettings(scenario = "CohisionAndCollision", nEnv = 1, nAgents = nAgents, nTargets = 8,
+        nSteps = 1000, nEpochs = 150, device = "cpu")
+    val environment = new VmasEnvironment(rewardFunction, actions, envSettings, WANDBLogger)
     var agents = Seq[VMASAgent]()
     for (i <- 0 until nAgents) {
         val agent = new VMASAgent(environment, actions, memory.asInstanceOf[ReplayBuffer[State, Action]])
@@ -30,8 +33,8 @@ class Main extends App {
         environment = environment,
         dataset = memory.asInstanceOf[ReplayBuffer[State, Action]],
         actionSpace = actions,
-        learningConfiguration = learningConfiguration
-    )(ExecutionContext.global)
+        learningConfiguration = learningConfiguration,
+        learner = Some(dql)
+    )(ExecutionContext.global, VMASState.encoding)
     ctde.learn(envSettings.nSteps, envSettings.nEpochs)
-
 }